Refactor to use common components and channels instead of waitgroup
