+++ jobs/import_process/worker_linux.go
@@ -7,125 +7,64 @@ package import_process
 import (
 	"archive/zip"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"syscall"
 
 	"github.com/mattermost/mattermost-server/v5/app"
 	"github.com/mattermost/mattermost-server/v5/model"
+	"github.com/mattermost/mattermost-server/v5/shared/filestore"
 	"github.com/mattermost/mattermost-server/v5/shared/mlog"
 )
 
-func (w *ImportProcessWorker) doJob(job *model.Job) {
-	if claimed, err := w.jobServer.ClaimJob(job); err != nil {
-		mlog.Warn("Worker experienced an error while trying to claim job",
-			mlog.String("worker", w.name),
-			mlog.String("job_id", job.Id),
-			mlog.String("error", err.Error()))
-		return
-	} else if !claimed {
-		return
-	}
-
-	importFileName, ok := job.Data["import_file"]
-	if !ok {
-		appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.missing_file", nil, "", http.StatusBadRequest)
-		w.setJobError(job, appError)
-		return
-	}
-
-	importFilePath := filepath.Join(*w.app.Config().ImportSettings.Directory, importFileName)
-	if ok, err := w.app.FileExists(importFilePath); err != nil {
-		w.setJobError(job, err)
-		return
-	} else if !ok {
-		appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, "", http.StatusBadRequest)
-		w.setJobError(job, appError)
-		return
-	}
-
-	importFileSize, appErr := w.app.FileSize(importFilePath)
-	if appErr != nil {
-		w.setJobError(job, appErr)
-		return
-	}
-
-	// create a temporary dir to extract the zipped import file.
-	dir, err := ioutil.TempDir("", "import")
-	if err != nil {
-		appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.tmp_dir", nil, err.Error(), http.StatusInternalServerError)
-		w.setJobError(job, appError)
-		return
-	}
-	defer os.RemoveAll(dir)
-
-	importFile, appErr := w.app.FileReader(importFilePath)
-	if appErr != nil {
-		w.setJobError(job, appErr)
-		return
-	}
-	defer importFile.Close()
-
+func (w *ImportProcessWorker) unzipAndImport(job *model.Job, unpackDirectory string, importFile filestore.ReadCloseSeeker, importFileSize int64, _ string) *model.AppError {
 	// extract the contents of the zipped file.
 	zipReader, err := zip.NewReader(importFile.(io.ReaderAt), importFileSize)
 	if err != nil {
 		appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-		w.setJobError(job, appError)
-		return
+		return appError
 	}
 
 	var jsonFile io.ReadCloser
-	var wg sync.WaitGroup
+	countFiles := len(zipReader.File)
+	errors := make(chan *model.AppError)
 	for _, zipFile := range zipReader.File {
 		if jsonFile == nil && filepath.Ext(zipFile.Name) == ".jsonl" {
 			jsonFile, err = zipFile.Open()
 			if err != nil {
-				appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-				w.setJobError(job, appError)
-				return
+				return model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
 			}
 			continue
 		}
 		zipFileName, err := filepath.Abs(zipFile.Name)
 		if err != nil {
 			appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-			w.setJobError(job, appError)
-			return
+			return appError
 		}
 		if strings.Contains(zipFileName, "..") {
-			appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, "illegal .. found in zipfile file path", http.StatusBadRequest)
-			w.setJobError(job, appError)
-			return
+			return model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, "illegal .. found in zipfile file path", http.StatusBadRequest)
 		}
-		namedPipePath := filepath.Join(dir, zipFileName)
+		namedPipePath := filepath.Join(unpackDirectory, zipFileName)
 		err = os.MkdirAll(filepath.Dir(namedPipePath), 0700)
 		if err != nil {
-			appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-			w.setJobError(job, appError)
-			return
+			return model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
 		}
 		mlog.Debug("Opening pipe", mlog.String("pipe", namedPipePath))
 		err = syscall.Mknod(namedPipePath, syscall.S_IFIFO|0666, 0)
 		if err != nil {
-			appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-			w.setJobError(job, appError)
-			return
+			return model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
 		}
 
-		go func(wg *sync.WaitGroup, zipFile *zip.File, namedPipePath string) {
+		go func(zipFile *zip.File, namedPipePath string, errors chan<- *model.AppError) {
 			mlog.Debug("Waiting for file to be read", mlog.String("pipe", namedPipePath))
-			wg.Add(1)
-			defer wg.Done()
 			namedPipe, err := os.OpenFile(namedPipePath, os.O_WRONLY, 0666)
 			if err != nil {
 				appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-				w.setJobError(job, appError)
+				errors <- appError
 				return
 			}
 			defer namedPipe.Close()
@@ -133,7 +72,7 @@ func (w *ImportProcessWorker) doJob(job *model.Job) {
 			fileAttachment, err := zipFile.Open()
 			if err != nil {
 				appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-				w.setJobError(job, appError)
+				errors <- appError
 				return
 			}
 
@@ -143,23 +82,27 @@ func (w *ImportProcessWorker) doJob(job *model.Job) {
 			_, err = io.Copy(namedPipe, fileAttachment)
 			if err != nil {
 				appError := model.NewAppError("ImportProcessWorker", "import_process.worker.do_job.file_exists", nil, err.Error(), http.StatusBadRequest)
-				w.setJobError(job, appError)
+				errors <- appError
 				return
 			}
 
+			errors <- nil
 			mlog.Debug("Done with pipe", mlog.String("pipe", namedPipePath))
-		}(&wg, zipFile, namedPipePath)
+		}(zipFile, namedPipePath, errors)
 	}
 
 	// do the actual import.
-	appErr, lineNumber := w.app.BulkImportWithPath(w.appContext, jsonFile, false, runtime.NumCPU(), filepath.Join(dir, app.ExportDataDir))
+	appErr, lineNumber := w.app.BulkImportWithPath(w.appContext, jsonFile, false, runtime.NumCPU(), filepath.Join(unpackDirectory, app.ExportDataDir))
 	if appErr != nil {
 		job.Data["line_number"] = strconv.Itoa(lineNumber)
-		w.setJobError(job, appErr)
-		return
+		return appErr
 	}
 
-	wg.Wait()
-	mlog.Info("Worker: Job is complete", mlog.String("worker", w.name), mlog.String("job_id", job.Id))
-	w.setJobSuccess(job)
+	for completed := 0; completed < countFiles; completed++ {
+		appErr := <-errors
+		if appErr != nil {
+			return appErr
+		}
+	}
+	return nil
 }
